+++ spy/tool/spyformatter.py
